no longer consider ScheduleTreeElemMappingFilter to be a subclass of filter
This type was introduced as filter because it is used as a filter
during AST generation.  However, AST generation is performed on
an isl schedule tree and a ScheduleTreeElemMappingFilter can easily
be converted to an isl filter independently of whether
ScheduleTreeElemMappingFilter itself is a filter.
Having ScheduleTreeElemMappingFilter as a subclass is causing
more problems than it is worth because every time a reference
is made to a filter, a careful consideration has to be made
whether or not to include the mapping filters.
Simply stop considering ScheduleTreeElemMappingFilter objects as filters
and deal with them specifically when needed.

In order to minimize the changes, the filter_ field is copied
into ScheduleTreeElemMappingFilter.  This filter_ contains
information that can be derived from the mapping, so it could
be replaced by a function that computes the filter later.