PolyhedralMapperTest.Match1: drop check for the number of parameters
This is a test for the tree matchers, so the number of parameters
is not really relevant.  isl_space_dim will not get exported
in mainline.  Perhaps some other mechanism will be introduced later
for obtaining the number of parameters, but this use case
is not sufficient motivation for introducing one.