MappedScop::makeOneBlockOneThread: return proper mapped scop
One of the invariants of a MappedScop is that there should
be mappings to blocks and threads in every path of the tree.
In particular, tightenLaunchBounds assumes that these mappings
are available, even though it will silently compute possibly
incorrect results (which happen to be correct on a scop
mapped with makeOneBlockOneThread) if they are missing.
Insert the required mappings in makeOneBlockOneThread.