Make number of elites less fragile
At the moment it is too easy to break autotuning by specifying a
pop_size that is smaller than the default number of elites.
Specifying the number of elites everywhere a small pop_size is
required is annoying and unnecessary cognitive overhead.

Instead, if one wants to run a short tuning run, default the number of
elites to `pop_size/2`, which also works for `3` the minimal size for
which a short tuning run makes sense.