Simplify and update TunerConfig
TunerConfig acts similarly to a python ContextManager (similar to RAII effects).
Therefore we can simplify the user-facing API and just construct a TunerConfig
instead of calling a bunch of methods that will set up a bunch of flags to
non-default values and never set them back.

Additionally, add the previously missing TunerConfig fields to the user-facing
API.