User-facing Python API
This commit reimplements the user-facing Python API and address numerous issues:
1. significant overhead on the python side, due in particular to caching with
python string keys based on input tensor sizes;
2. significant overhead on the tensor allocation side because of multiple
transformations on the ATen to TensorInfo path + range inferencing every time;
3. implicit behaviors everywhere which gave a distorted view of how to use TC
properly (see issue #196). This is also contrary to PEP-20;
4. 1 cache file per TC and size combination which makes it very cumbersome
to use caching to files;
5. every python side API call being implemented in terms of
`TCFunction.apply(self.cu, tc_info, kwargs, *inputs)` which puts the PyTorch
python autograd abstraction on the critical path to everything. This is not
portable or even future-proof since with PyTorch 0.5 we want to integrate TC
at the C++ level after measuring 100+us for just getting into a torch.backward;
6. benchmark overhead of the abstraction which was quite more expensive than
the CUDA kernels themselves in low latency mode;
7. `tc.define` abstraction only allows 1 TC function for `forward` and 1 TC
function for `backward` and has a very clunky interface with a function to map
tensors to the inputs of `backward`

After design discussions with @apaszke it became quickly clear that
using python language features was a significantly better way to go.
This rewrite also kills overhead everywhere and removes as much implicit
behavior as possible.

A little implicit behavior is still remaining in the form of how MappingOptions
are generated in TC class function calls on the first compilation (i.e. similar
behavior to cudnn benchmark except here we JIT-compile).
To make that behavior as explicit as possible we require the user to pass a
generating function which returns a MappingOptions object.
We provide 3 such functions:
1. always naive options for debug mode
2. load from cache or fail for using with a compilation cache file
3. autotune backed by cache with the same API as the `tc.autotune` function

In the future we can add more behaviors depending on user feedback.
For now it seems safer to let the user use whatever preferred behavior.