ScheduleTree: copy underlying element in makeScheduleTree
A subsequent commit is going to replace schedule trees containing
elements by elements inheriting the tree, removing the "elem_" field.
Since the copy constructor of the tree is not intended to be called
otherwise than through a call to makeScheduleTree, move the copy of
underlying element to that function.  It will allow the subsequent
commit to call the dedicated static constructors instead.