ScheduleTreeElemMapping: take isl::ctx in constructor
The constructor of ScheduleTreeElemMapping accepts as argument an
unordered_map between isl objects, so the context may be extracted from
any object unless the map is empty.  While the constructor is throwing
if the map is empty, the isl::ctx instance is passed to the superclass
constructor, that is before it has an opportunity to throw.  Request the
context as a separate constructor argument instead of hoping on a
non-empty map and potentially triggering an undefined behavior by
dereferencing the one-past-end iterator.